@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2017 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT licence. See License.txt in the project root for license information.
 
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 
 namespace TestSupport.EfHelpers
 {
