@@ -45,7 +45,8 @@ namespace Test.UnitTests.EfSchemaCompare
                 var dtService = context.GetDesignTimeService();
                 var serviceProvider = dtService.GetDesignTimeProvider();
                 var factory = serviceProvider.GetService<IDatabaseModelFactory>();
-                var database = factory.Create(_connectionString, new string[] { }, new string[] { });
+                var database = factory.Create(_connectionString,  
+                    new DatabaseModelFactoryOptions(new string[] { }, new string[] { }));
                 var handler = new Stage1Comparer(context.Model, nameof(BookOrderSchemaContext));
 
                 //ATTEMPT
