@@ -3,8 +3,8 @@
 
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.Reflection;
+using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Configuration;
 using TestSupport.Helpers;
 
