@@ -2,9 +2,9 @@
 // Licensed under MIT licence. See License.txt in the project root for license information.
 
 using System;
-using System.Data.SqlClient;
 using System.IO;
 using System.Reflection;
+using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Configuration;
 
 namespace TestSupport.Helpers
